pyproject.toml
[project]
name = "kitos-integration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flatten-json>=0.1.14",
    "ipykernel>=6.29.5",
    "pandas>=2.2.3",
    "pyodbc>=5.2.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.41",
]

[tool.ruff]
line-length = 120
